+++ modules/node-groups/variables.tf
@@ -28,8 +28,8 @@ variable "cluster_name" {
   nullable    = false
 }
 
-variable "cluster_version" {
-  description = "The AKS Kubernetes version."
+variable "cluster_version_full" {
+  description = "The full Kubernetes version of the Azure Kubernetes managed cluster."
   type        = string
   nullable    = false
 }